+++ fast/stages/1-resman/branch-networking.tf
@@ -68,6 +68,7 @@ module "branch-network-prod-folder" {
       local.branch_optional_sa_lists.dp-prod,
       local.branch_optional_sa_lists.gke-prod,
       local.branch_optional_sa_lists.gcve-prod,
+      local.branch_optional_sa_lists.pf,
       local.branch_optional_sa_lists.pf-prod,
     )
     # read-only (plan) automation service accounts
@@ -75,6 +76,7 @@ module "branch-network-prod-folder" {
       local.branch_optional_r_sa_lists.dp-prod,
       local.branch_optional_r_sa_lists.gke-prod,
       local.branch_optional_r_sa_lists.gcve-prod,
+      local.branch_optional_r_sa_lists.pf,
       local.branch_optional_r_sa_lists.pf-prod,
     )
     (local.custom_roles.gcve_network_admin) = local.branch_optional_sa_lists.gcve-prod
@@ -97,6 +99,7 @@ module "branch-network-dev-folder" {
       local.branch_optional_sa_lists.dp-dev,
       local.branch_optional_sa_lists.gke-dev,
       local.branch_optional_sa_lists.gcve-dev,
+      local.branch_optional_sa_lists.pf,
       local.branch_optional_sa_lists.pf-dev,
     )
     # read-only (plan) automation service accounts
@@ -104,6 +107,7 @@ module "branch-network-dev-folder" {
       local.branch_optional_r_sa_lists.dp-dev,
       local.branch_optional_r_sa_lists.gke-dev,
       local.branch_optional_r_sa_lists.gcve-dev,
+      local.branch_optional_r_sa_lists.pf,
       local.branch_optional_r_sa_lists.pf-dev,
     )
     (local.custom_roles.gcve_network_admin) = local.branch_optional_sa_lists.gcve-dev