+++ fast/stages/2-networking-a-simple/variables-fast.tf
@@ -94,6 +94,7 @@ variable "service_accounts" {
     data-platform-prod   = string
     gke-dev              = string
     gke-prod             = string
+    project-factory      = string
     project-factory-dev  = string
     project-factory-prod = string
   })