+++ fast/stages/2-security/core-prod.tf
@@ -18,6 +18,7 @@ locals {
   prod_kms_restricted_admins = [
     for sa in distinct(compact([
       var.service_accounts.data-platform-prod,
+      var.service_accounts.project-factory,
       var.service_accounts.project-factory-prod
     ])) : "serviceAccount:${sa}"
   ]